requirements.txt

flask
python-dotenv
line-bot-sdk
schedule
feedparser
langchain
langchain-core
langchain-openai


